Renamed CompanyProfileDao to CompanyProfileDocument, moved to model
