+++ src/main/java/uk/gov/companieshouse/company/profile/repository/CompanyProfileRepository.java
@@ -2,10 +2,10 @@ package uk.gov.companieshouse.company.profile.repository;
 
 import java.util.Optional;
 import org.springframework.data.mongodb.repository.MongoRepository;
-import uk.gov.companieshouse.company.profile.domain.CompanyProfileDao;
+import uk.gov.companieshouse.company.profile.model.CompanyProfileDocument;
 
-public interface CompanyProfileRepository extends MongoRepository<CompanyProfileDao, String> {
+public interface CompanyProfileRepository extends MongoRepository<CompanyProfileDocument, String> {
     // company number is not a direct property of the entity so traversal is required
-    Optional<CompanyProfileDao> findCompanyProfileDaoByCompanyProfile_Data_CompanyNumber(
+    Optional<CompanyProfileDocument> findCompanyProfileDaoByCompanyProfile_Data_CompanyNumber(
             String companyNumber);
 }
